chore: Add AppException class and ErrorCode enum for exception handling

+++ src/main/java/com/example/JewelrySalesSystem/service/UserService.java
@@ -3,6 +3,8 @@ package com.example.JewelrySalesSystem.service;
 import com.example.JewelrySalesSystem.dto.request.UserCreationRequest;
 import com.example.JewelrySalesSystem.dto.request.UserUpdateRequest;
 import com.example.JewelrySalesSystem.entity.User;
+import com.example.JewelrySalesSystem.exception.AppException;
+import com.example.JewelrySalesSystem.exception.ErrorCode;
 import com.example.JewelrySalesSystem.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,13 +16,12 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public User createUser(UserCreationRequest request){
-        User user = new User();
-
-        if (userRepository.existsByUsername(request.getUsername())){
-            throw new RuntimeException("Username already exists");
+    public User createUser(UserCreationRequest request) {
+        if (userRepository.existsByUsername(request.getUsername())) {
+            throw new AppException(ErrorCode.USER_EXISTED);
         }
 
+        User user = new User();
         user.setUsername(request.getUsername());
         user.setPassword(request.getPassword());
         user.setEmail(request.getEmail());
@@ -29,19 +30,18 @@ public class UserService {
         user.setRewardPoints(request.getRewardPoints());
 
         return userRepository.save(user);
-
     }
 
-    public List<User> getUsers(){
+    public List<User> getUsers() {
         return userRepository.findAll();
     }
 
-    public User getUser(String id){
+    public User getUser(String id) {
         return userRepository.findById(id)
-                .orElseThrow(()->new RuntimeException("User Not Found"));
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
     }
 
-    public User updateUser(String userID, UserUpdateRequest request){
+    public User updateUser(String userID, UserUpdateRequest request) {
         User user = getUser(userID);
 
         user.setPassword(request.getPassword());
@@ -53,7 +53,10 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(String userId){
+    public void deleteUser(String userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
+        }
         userRepository.deleteById(userId);
     }
 }
